requirements.txt
streamlit
python-dotenv
couchdb3
jsonschema
bcrypt
pytest-mock
pytest-cov